In Compose, there are multiple ways to animate your UI: from high-level APIs for simple animations to low-level methods for full control and complex transitions. You can read about them in the documentation.
In this section you will use one of the low-level APIs but don't worry, they can also be very simple. Let's animate the change in size that we already implemented:

For this you'll use the animateDpAsState composable. It returns a State object whose value will continuously be updated by the animation until it finishes. It takes a "target value" whose type is Dp