style any of the composables so far and yet you got a decent default, including dark mode support! Let's look into what BasicsCodelabTheme and MaterialTheme are.
If you open the ui/theme/Theme.kt file

MaterialTheme is a composable function that reflects the styling principles from the Material design specification. That styling information cascades down to the components that are inside its content, which may read the information to style themselves. In your UI, you are already using BasicsCodelabTheme
The Text composable in the example above sets a new TextStyle. You can create your own TextStyle, or you can retrieve a theme-defined style by using MaterialTheme.typography, which is preferred. This construct gives you access to the Material-defined text styles, such as displayLarge, headlineMedium, titleSmall, bodyLarge, labelMedium etc. In your example, you use the headlineMedium style defined in the theme.

You can find everything related to the current theme in the files inside the ui/theme folder. For example, the default colors that we have been using so far are defined in Color.kt.